chore: Try to add formatting rules identical to existing source
The additions were mostly produced by letting Rider infer the rules based on the existing source; the remaining few were added by reformatting files randomly and checking the diff (if any).
We use this in order to avoid lots of false-positives when looking at diffs.
Isn't perfect but the remaining rules are really hard to guess (there's a LOT of them).