feat: First instruction's test cases all pass
Added support for `v128.const` and due to the test cases's reliance on `i32x4.add` and `i64x2.add` had to add those too.
For now, only adding the `.wast` test file which can be used to regenerate the other files necessary for unit tests. TODO: add script to regenerate files and/or instructions on how to do this.