fix: Ensure that **all** tests pass, not just newly added ones
There was one regression which caused many _other_ tests to fail: when reading the element section, we only need to read additional uint32 when the element section kind is 2 (this kind was added recently). The original code somewhat missed the spec and misrepresented the first uint32.

Then, there were several "stylistic" tests which ensured uniform naming and visibility of the instructions, made them pass too.