feat: Almost implemented tests for simd bitwise operations
The last section uses the `v128.load` which requires a bit more "machinery" to function: can't derive from `MemoryImmediateInstruction` because of the disparity between its base class `Instruction` and `SimdInstruction`. Though we now can use 16 byte alignment and helper methods are static so can be reused. Some minor copy-paste and we should be good for most `v128.load` variations.