feat: Implement store and all store lane variations
The key obstacle was the memory access prolog helper which, as its last instruction, emitted `Unaligned` which is required to be followed by a memory access instruction.
This is always the case except in the "store lane" case where we need a pointer _and_ the value to store so we slightly rearranged code to do this correctly.
Once it was clear what the invalid IL was, the fix/workaround was easy but it was **not at all** easy to figure this out (AI 'guessed it' in the end, after a lot of experimentation).